requirements.txt
Django==1.10.5
django-ace-overlay==0.6
